LayoutTests/fast/css/modify-ua-rules-from-javascript-expected.txt

CONSOLE MESSAGE: line 11: TypeError: Result of expression 'window.getMatchedCSSRules(document.body, "", false)' [null] is not an object.
PASS
